Core: move Unit to an independent class
