+++ src/main/java/org/semux/api/util/TransactionBuilder.java
@@ -6,7 +6,7 @@
  */
 package org.semux.api.util;
 
-import static org.semux.core.Amount.Unit.NANO_SEM;
+import static org.semux.core.Unit.NANO_SEM;
 
 import org.semux.Kernel;
 import org.semux.Network;
