+++ src/main/java/org/semux/config/AbstractConfig.java
@@ -6,10 +6,10 @@
  */
 package org.semux.config;
 
-import static org.semux.core.Amount.Unit.MILLI_SEM;
-import static org.semux.core.Amount.Unit.NANO_SEM;
-import static org.semux.core.Amount.Unit.SEM;
 import static org.semux.core.Amount.ZERO;
+import static org.semux.core.Unit.MILLI_SEM;
+import static org.semux.core.Unit.NANO_SEM;
+import static org.semux.core.Unit.SEM;
 
 import java.io.File;
 import java.io.FileInputStream;
