+++ src/main/java/org/semux/core/state/DelegateStateImpl.java
@@ -6,10 +6,10 @@
  */
 package org.semux.core.state;
 
-import static org.semux.core.Amount.Unit.NANO_SEM;
 import static org.semux.core.Amount.ZERO;
 import static org.semux.core.Amount.sub;
 import static org.semux.core.Amount.sum;
+import static org.semux.core.Unit.NANO_SEM;
 
 import java.util.ArrayList;
 import java.util.Arrays;
