+++ src/main/java/org/semux/gui/SwingUtil.java
@@ -8,7 +8,7 @@ package org.semux.gui;
 
 import static java.util.Arrays.stream;
 
-import static org.semux.core.Amount.Unit.SEM;
+import static org.semux.core.Unit.SEM;
 import static org.semux.gui.TextContextMenuItem.COPY;
 import static org.semux.gui.TextContextMenuItem.CUT;
 import static org.semux.gui.TextContextMenuItem.PASTE;
@@ -58,8 +58,8 @@ import javax.swing.table.TableColumnModel;
 
 import org.ethereum.vm.util.HashUtil;
 import org.semux.core.Amount;
-import org.semux.core.Amount.Unit;
 import org.semux.core.Transaction;
+import org.semux.core.Unit;
 import org.semux.core.state.Delegate;
 import org.semux.core.state.DelegateState;
 import org.semux.crypto.Hex;
