+++ src/test/java/org/semux/bench/BlockchainPerformance.java
@@ -6,7 +6,7 @@
  */
 package org.semux.bench;
 
-import static org.semux.core.Amount.Unit.NANO_SEM;
+import static org.semux.core.Unit.NANO_SEM;
 
 import java.util.ArrayList;
 import java.util.List;
