+++ src/test/java/org/semux/bench/VMPerformance.java
@@ -6,8 +6,8 @@
  */
 package org.semux.bench;
 
-import static org.semux.core.Amount.Unit.NANO_SEM;
-import static org.semux.core.Amount.Unit.SEM;
+import static org.semux.core.Unit.NANO_SEM;
+import static org.semux.core.Unit.SEM;
 
 import java.util.ArrayList;
 import java.util.Collections;
