+++ src/test/java/org/semux/config/MainnetConfigTest.java
@@ -11,8 +11,8 @@ import static org.hamcrest.Matchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
-import static org.semux.core.Amount.Unit.SEM;
 import static org.semux.core.Amount.ZERO;
+import static org.semux.core.Unit.SEM;
 
 import java.io.File;
 import java.io.IOException;
