+++ src/test/java/org/semux/gui/dialog/TransactionDialogTest.java
@@ -6,8 +6,8 @@
  */
 package org.semux.gui.dialog;
 
-import static org.semux.core.Amount.Unit.SEM;
 import static org.semux.core.TransactionType.TRANSFER;
+import static org.semux.core.Unit.SEM;
 
 import java.math.BigDecimal;
 import java.time.Instant;
