+++ src/test/java/org/semux/gui/panel/DelegatesPanelTest.java
@@ -13,8 +13,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.semux.core.Amount.Unit.SEM;
 import static org.semux.core.Amount.ZERO;
+import static org.semux.core.Unit.SEM;
 
 import java.util.ArrayList;
 import java.util.List;
